% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppindices-package.R
\docType{package}
\name{ppindices-package}
\alias{ppindices}
\alias{ppindices-package}
\title{ppindices: Phenotypic Plasticity Indices}
\description{
A collection of implementations for quantifying phenotypic plasticity from
reaction-norm data. Provides functions to compute a broad range of
plasticity indices from trait values measured across environments.
}
\author{
\strong{Maintainer}: Benedikt Haarscheidt \email{benedikt.haarscheidt@yahoo.de}

}
\keyword{internal}
